requirements.txt
python-telegram-bot
requests
beautifulsoup4 